// src/models/Zona.ts
/**
 *
 * @export
 * @interface Zona
 */
export interface Zona {
      /**
     *
     * @type {number}
     * @memberof Zona
     */
      id: number;
    /**
     *
     * @type {string}
     * @memberof Zona
     */
    nombre: string;
    /**
     *
     * @type {string}
     * @memberof Zona
     */
    descripcion: string;
    
  }
  